Remove lib/client/ipam subpackage
