+++ lib/client/ipam_block.go
@@ -1,4 +1,4 @@
-package ipam
+package client
 
 import (
 	"errors"
