+++ lib/client/ipam_types.go
@@ -1,0 +1,80 @@
+// Copyright (c) 2016 Tigera, Inc. All rights reserved.
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package client
+
+import (
+	"net"
+)
+
+// AssignIPArgs defines the set of arguments for assigning a specific IP address.
+type AssignIPArgs struct {
+	// The IP address to assign.
+	IP net.IP
+
+	// If specified, a handle which can be used to retrieve / release
+	// the allocated IP addresses in the future.
+	HandleID *string
+
+	// A key/value mapping of metadata to store with the allocations.
+	Attrs map[string]string
+
+	// If specified, the hostname of the host on which IP addresses
+	// will be allocated.  If not specified, this will default
+	// to the value provided by os.Hostname.
+	Hostname *string
+}
+
+// AutoAssignArgs defines the set of arguments for assigning one or more
+// IP addresses.
+type AutoAssignArgs struct {
+	// The number of IPv4 addresses to automatically assign.
+	Num4 int
+
+	// The number of IPv6 addresses to automatically assign.
+	Num6 int
+
+	// If specified, a handle which can be used to retrieve / release
+	// the allocated IP addresses in the future.
+	HandleID *string
+
+	// A key/value mapping of metadata to store with the allocations.
+	Attrs map[string]string
+
+	// If specified, the hostname of the host on which IP addresses
+	// will be allocated.  If not specified, this will default
+	// to the value provided by os.Hostname.
+	Hostname *string
+
+	// If specified, the previously configured IPv4 pool from which
+	// to assign IPv4 addresses.  If not specified, this defaults to all IPv4 pools.
+	IPv4Pool *net.IPNet
+
+	// If specified, the previously configured IPv6 pool from which
+	// to assign IPv6 addresses.  If not specified, this defaults to all IPv6 pools.
+	IPv6Pool *net.IPNet
+}
+
+// IPAMConfig contains global configuration options for Calico IPAM.
+type IPAMConfig struct {
+	// When StrictAffinity is true, addresses from a given block can only be
+	// assigned by hosts with the blocks affinity.  If false, then AutoAllocateBlocks
+	// must be true.
+	StrictAffinity bool
+
+	// When AutoAllocateBlocks is true, the IPAM client will automatically
+	// allocate blocks as needed to assign addresses.  If false, then
+	// StrictAffinity must be true.
+	AutoAllocateBlocks bool
+}
